Moved files to experimenter and fixed logging mishaps
